+++ Sources/VCore/Services and Managers/Modal Presenter/Root/Overlay/ModalPresenterRootViewModifier_Overlay.swift
@@ -228,8 +228,6 @@ struct ModalPresenterRootViewModifier_Overlay: ViewModifier {
 
         modal.presentationMode.dismissSubject.send /*completion: */{
             modals.removeAll { $0.id == dismissData.link.linkID }
-            
-            ModalPresenterDataSourceCache.shared.remove(link: dismissData.link)
 
             dismissData.completion()
         }